backend/requirements.txt
fastapi
google-generativeai
httpx
python-dotenv
uvicorn[standard]
